Use a global variable for current test
This eliminates the need for lambdas to capture, which was very
annoying and easy to forget.

The original reason for passing current test into the test case was to
anticipate making the library multi-threaded. However, there are no
plans to do that at the moment, so it's better to make current usage
easier and refactor later if needed.

It also means the THIS_TEST macro can be replaced with a function,
and the THIS_TEST_AS macro with a function template.

Also fixed a typo in the docs.